Us 565 RG for local, sandbox and demo/qa env (#16)
* Removed unnecessary code and files

* Refactored code

* Modified Readme

* Modified code run RG in sandbox env

* Fixed request and azureml-core package version conflit

* Fixed posting sample files to sandbox env

* Fixed issue which fails due to pip

* Fixed flake8 error

* Added env variables for authentication

* pepify code

* Fixed issue for format type of rgb

* Added endpoint url for Demo environment

* Fixed env issue and made changes json load suggestion

* Added demo url to api endpoints

* Added missing code for DEMO env

* pepify the code

* Remove unnecessary code and renamed files

* Pepify the code and renamed the file
