--- entrypoint.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-printenv | grep -v "no_proxy" >> /etc/environment
-
-cron -f
