+++ requirements.txt
@@ -17,3 +17,4 @@ cmake
 dlib
 face-recognition==1.3.0
 jsonschema==2.6.0
+requests==2.19.1
