--- src/run_env.py
@@ -1,1 +1,0 @@
-RUN_ENV = ''
